-created GoogleFitDataImporter service -moved ImportFitnessData method to new created service
