+++ src/GoogleFitOnFhir/Services/IGoogleFitDataImporter.cs
@@ -1,18 +1,15 @@
-// -------------------------------------------------------------------------------------------------
+﻿// -------------------------------------------------------------------------------------------------
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 // -------------------------------------------------------------------------------------------------
 
 using System.Threading;
 using System.Threading.Tasks;
-using GoogleFitOnFhir.Models;
 
 namespace GoogleFitOnFhir.Services
 {
-    public interface IUsersService
+    public interface IGoogleFitDataImporter
     {
-        Task<User> Initiate(string accessCode, CancellationToken cancellationToken);
-
-        void QueueFitnessImport(User user);
+        Task ImportFitnessData(string userId, CancellationToken cancellationToken);
     }
 }
